Add files via upload
These images include the image box that we have created (can be viewed in the solidworks file) as well as the LFA strips with their respective samples. There is an unknown sample that we used to test and observe whether the glucose levels match up with the samples that we have measured.